refactor: cleanup history meta and composable abstraction
