+++ frontend/src/composables/useTextEditor.js
@@ -2,7 +2,7 @@ import { ref, reactive, watch } from 'vue'
 import { Editor } from '@tiptap/vue-3'
 import { extensions } from '@/stores/tiptapSetup'
 import { TextSelection } from 'prosemirror-state'
-import { commandHistory } from '@/stores/history'
+import { commandHistory } from '@/stores/historyMeta'
 import { activeElement } from '@/stores/element'
 import { editElementCommand } from '@/stores/commands'
 import { currentSlide } from '@/stores/slide'
