--- frontend/src/stores/history.js
@@ -1,7 +1,0 @@
-let commandHistory = null
-
-const setCommandHistory = (history) => {
-	commandHistory = history
-}
-
-export { setCommandHistory, commandHistory }
