+++ frontend/src/stores/presentation.js
@@ -6,7 +6,7 @@ import { router } from '@/router'
 import { slides } from './slide'
 import { normalizeZIndices } from '@/stores/element'
 import { v4 as uuid4 } from 'uuid'
-import { commandHistory } from './history'
+import { commandHistory } from './historyMeta'
 
 const isDriveInstalled = window.apps?.includes('drive') ?? false
